requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.35.0
billiard==4.2.1
botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
boto==2.49.0
boto3==1.40.6
botocore==1.40.6
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
Django==4.2.23
django-widget-tweaks==1.5.0
frozenlist==1.7.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
jsonpickle==1.4.2
kombu==5.5.4
lxml==6.0.0
msal==1.33.0
msrest==0.7.1
multidict==6.6.4
oauthlib==3.3.1
packaging==25.0
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
pycparser==2.22
PyJWT==2.10.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
redis==6.4.0
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.13.1
six==1.17.0
slack_sdk==3.36.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
